+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-## 1.2.0
+## 1.3.0
+* The big update
+* Improved semantics
+* Added simple diagnostic provider 
+* Improved hovers
+* Symbol documenter is faster
+
+## 1.2.0 (Unreleased)
 * Added Completion Provider
 * Added ez80 snippets and documentation
 * Added settings regarding ez80 instruction snippets